+++ src/config/tokenIdentifiers.ts
@@ -3,37 +3,29 @@ import { TokenIdentifierMap } from './types'
 // --- TESTNET TOKEN IDENTIFIERS ---
 // IMPORTANT make name same as tokens[token].symbol value
 export const testnetTokenIdentifiers: TokenIdentifierMap = {
-  WETH: {
+  rWETH: {
     domain: 'rinkeby',
     id: '0xc778417e063141139fce010982780140aa0cd5ab',
-  },
-  kWETH: {
-    domain: 'kovan',
-    id: '0xd0a1e359811322d97991e03f863a0c30c2cf029c',
   },
   gWETH: {
     domain: 'goerli',
     id: '0x0B1ba0af832d7C05fD64161E0Db78E85978E8082',
   },
   WBTC: {
-    domain: 'kovan',
-    id: '0xe0f131fb595000d7e54049efe5c40dca9572469c',
+    domain: 'goerli',
+    id: '0xaDcf53aD759054Dee4139d7aD206cC11788A466D',
   },
   USDC: {
+    domain: 'goerli',
+    id: '0x07865c6E87B9F70255377e024ace6630C1Eaa37F',
+  },
+  rUSDC: {
     domain: 'rinkeby',
     id: '0x4dbcdf9b62e891a7cec5a2568c3f4faf9e8abe2b',
   },
-  USDT: {
-    domain: 'kovan',
-    id: '0x13512979ade267ab5100878e2e0f485b568328a4',
-  },
-  DAI: {
-    domain: 'kovan',
-    id: '0xFf795577d9AC8bD7D90Ee22b6C1703490b6512FD',
-  },
   TEST: {
-    domain: 'kovan',
-    id: '0xe71678794fff8846bff855f716b0ce9d9a78e844',
+    domain: 'goerli',
+    id: '0x8a1Cad3703E0beAe0e0237369B4fcD04228d1682',
   },
   WXDAI: {
     domain: 'xdai',
@@ -42,6 +34,10 @@ export const testnetTokenIdentifiers: TokenIdentifierMap = {
   wtEVMOS: {
     domain: 'evmostestnet',
     id: '0xEE8003bdB024a4195D433c555d55AA663B8C022C',
+  },
+  wtNEON: {
+    domain: 'neontestnet',
+    id: '0xf8ad328e98f85fccbf09e43b16dcbbda7e84beab',
   },
   //   DEV: {
   //     domain: 'moonbasealpha',
@@ -59,6 +55,18 @@ export const mainnetTokenIdentifiers: TokenIdentifierMap = {
   WETH: {
     domain: 'ethereum',
     id: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+  },
+  wEVMOS: {
+    domain: 'evmos',
+    id: '0xD4949664cD82660AaE99bEdc034a0deA8A0bd517',
+  },
+  WADA: {
+    domain: 'milkomedaC1',
+    id: '0xAE83571000aF4499798d1e3b0fA0070EB3A3E3F9',
+  },
+  WGLMR: {
+    domain: 'moonbeam',
+    id: '0xAcc15dC74880C9944775448304B263D191c6077F',
   },
   WBTC: {
     domain: 'ethereum',
@@ -84,14 +92,12 @@ export const mainnetTokenIdentifiers: TokenIdentifierMap = {
     domain: 'ethereum',
     id: '0x3432B6A60D23Ca0dFCa7761B7ab56459D9C964D0',
   },
-  WGLMR: {
-    domain: 'moonbeam',
-    id: '0xAcc15dC74880C9944775448304B263D191c6077F',
+  MIM: {
+    domain: 'ethereum',
+    id: '0x99d8a9c45b2eca8864373a26d1459e3dff1e17f3',
   },
-  WADA: {
-    domain: 'milkomedaC1',
-    id: '0xAE83571000aF4499798d1e3b0fA0070EB3A3E3F9',
-  },
+
+  // others
   WSTR: {
     domain: 'ethereum',
     id: '0xf0dc76c22139ab22618ddfb498be1283254612b1',
@@ -99,10 +105,6 @@ export const mainnetTokenIdentifiers: TokenIdentifierMap = {
   CQT: {
     domain: 'ethereum',
     id: '0xD417144312DbF50465b1C641d016962017Ef6240',
-  },
-  wEVMOS: {
-    domain: 'evmos',
-    id: '0xD4949664cD82660AaE99bEdc034a0deA8A0bd517',
   },
   CARDS: {
     domain: 'ethereum',
@@ -132,6 +134,42 @@ export const mainnetTokenIdentifiers: TokenIdentifierMap = {
     domain: 'ethereum',
     id: '0x0316EB71485b0Ab14103307bf65a021042c6d380',
   },
+  CRO: {
+    domain: 'ethereum',
+    id: '0xa0b73e1ff0b80914ab6fe0444e65848c4c34450b',
+  },
+  UNI: {
+    domain: 'ethereum',
+    id: '0x1f9840a85d5af5bf1d1762f925bdaddc4201f984',
+  },
+  SHIB: {
+    domain: 'ethereum',
+    id: '0x95ad61b0a150d79219dcf64e1e6cc01f0b64c4ce',
+  },
+  LINK: {
+    domain: 'ethereum',
+    id: '0x514910771AF9Ca656af840dff83E8264EcF986CA',
+  },
+  APE: {
+    domain: 'ethereum',
+    id: '0x40e0a6ef9dbadfc83c5e0d15262feb4638588d77',
+  },
+  MKR: {
+    domain: 'ethereum',
+    id: '0x9f8F72aA9304c8B593d555F12eF6589cC3A579A2',
+  },
+  COMP: {
+    domain: 'ethereum',
+    id: '0xc00e94cb662c3520282e6f5717214004a7f26888',
+  },
+  AAVE: {
+    domain: 'ethereum',
+    id: '0x7fc66500c84a76ad7e9c93437bfc5ac33e2ddae9',
+  },
+  TIC: {
+    domain: 'avalanche',
+    id: '0x75739a693459f33b1fbcc02099eea3ebcf150cbe',
+  },
   HBOT: {
     domain: 'ethereum',
     id: '0xE5097D9baeAFB89f9bcB78C9290d545dB5f9e9CB',
@@ -143,5 +181,9 @@ export const mainnetTokenIdentifiers: TokenIdentifierMap = {
   PBX: {
     domain: 'ethereum',
     id: '0xd528cf2e081f72908e086f8800977df826b5a483',
+  },
+  GEL: {
+    domain: 'ethereum',
+    id: '0x15b7c0c907e4c6b9adaaaabc300c08991d6cea05',
   },
 }