+++ src/utils/index.ts
@@ -2,6 +2,7 @@ import { BigNumber, utils } from 'ethers'
 
 import { networks, tokens } from '@/config'
 import { NetworkMetadata, TokenMetadata, NetworkName } from '@/config/types'
+import { TokenIdentifier } from '@nomad-xyz/sdk-bridge'
 
 const coinGeckoIds = Object.values(tokens).map((t) => t.coinGeckoId)
 
@@ -55,15 +56,23 @@ export function truncateAddr(addr: string): string {
 }
 
 export function fromBytes32(addr: string): string {
-  // trim 12 bytes from beginning plus '0x'
-  const short = addr.slice(26)
-  return `0x${short}`
+  if (addr.length === 42) return addr
+  if (addr.length === 66) {
+    // trim 12 bytes from beginning plus '0x'
+    const short = addr.slice(26)
+    return `0x${short}`
+  }
+  throw new Error('invalid address length, cannot convert to 20 bytes')
 }
 
 export function toBytes32(addr: string): string {
-  // trim 12 bytes from beginning plus '0x'
-  const short = addr.slice(2)
-  return `0x000000000000000000000000${short}`
+  if (addr.length === 66) return addr
+  if (addr.length === 42) {
+    // trim 12 bytes from beginning plus '0x'
+    const short = addr.slice(2)
+    return `0x000000000000000000000000${short}`
+  }
+  throw new Error('invalid address length, cannot convert to 20 bytes')
 }
 
 /**
@@ -206,6 +215,7 @@ export function getNetworkDomainIDByName(networkName: string): number {
 // TOKEN
 
 export const nullToken: TokenMetadata = {
+  key: '',
   nativeNetwork: 'ethereum',
   symbol: '',
   name: '',
@@ -214,13 +224,28 @@ export const nullToken: TokenMetadata = {
   coinGeckoId: '',
   tokenIdentifier: null,
   nativeOnly: false,
-  minAmt: 0,
 }
 
 export function getTokenBySymbol(symbol: string): TokenMetadata {
   for (const t in tokens) {
     const token = tokens[t]
     if (token.symbol === symbol) {
+      return token
+    }
+  }
+  throw new Error('token asset not found')
+}
+
+export function getTokenByTokenID(
+  tokenID: TokenIdentifier
+): TokenMetadata | undefined {
+  for (const t in tokens) {
+    const token = tokens[t]
+    if (
+      token.tokenIdentifier &&
+      token.tokenIdentifier.id.toString().toLowerCase() ==
+        tokenID.id.toString().toLowerCase()
+    ) {
       return token
     }
   }