+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This change log is manually updated at the moment.
 
 ## Unreleased
 
+## v1.8.0 (July 14, 2022)
+
+- transaction history
+- token search
+
 ## v1.7.1 (June 23, 2022)
 
 - feature: fetch hosted config
@@ -12,6 +17,15 @@ This change log is manually updated at the moment.
 ## v1.6.1 (June 23, 2022)
 
 - add: link to tx search in nav
+
+## v1.5.3 (June 13th, 2022)
+
+- sentry user id (wallet address)
+- "Add token" upgrade
+- search token bar
+- basic segment page tracking
+- fix missing avalanche icon
+- feat: segment page tracking
 
 ## v1.6.0 (June 22, 2022)